+++ components/BillingPlans.vue
@@ -4,15 +4,17 @@
 		<strong>Plans</strong>
 		<ul v-if="currency !== null && plans.length">
 			<li v-for="plan in plans" :key="plan.name">
-				<strong class="plan-name">
-					<a v-if="plan.url" :href="plan.url" target="_blank">{{ plan.name }}</a>
-					<template v-else>{{ plan.name }}</template>
-				</strong>
-				<ul class="badges small inline">
-					<li class="badge default" v-if="defaultPlan == plan.name">default</li>
-					<li class="badge red" v-if="plan.paid === true">paid</li>
-					<li class="badge green" v-else-if="plan.paid === false">free</li>
-				</ul>
+				<div class="plan-header">
+					<strong class="plan-name">
+						<a v-if="plan.url" :href="plan.url" target="_blank">{{ plan.name }}</a>
+						<template v-else>{{ plan.name }}</template>
+					</strong>
+					<ul class="badges small inline">
+						<li class="badge default" v-if="defaultPlan == plan.name">default</li>
+						<li class="badge red" v-if="plan.paid === true">paid</li>
+						<li class="badge green" v-else-if="plan.paid === false">free</li>
+					</ul>
+				</div>
 				<Description v-if="plan.description" :description="plan.description" :compact="true" />
 			</li>
 		</ul>
@@ -73,11 +75,14 @@ export default {
 		margin-bottom: 0.5em;
 	}
 
-	.plan-name {
-		margin: 0.25em 0;
-		padding: 0.2em 0;
-		line-height: 1em;
-		vertical-align: baseline;
+	.plan-header {
+		margin: 0.2em 0;
+
+		> .plan-name {
+			display: inline-block;
+			vertical-align: bottom;
+		}
 	}
+
 }
 </style>